+++ dev/dev-checks.R
@@ -11,11 +11,13 @@ covr::report()
 
 lintr::lint_package()
 goodpractice::gp(
-  checks = setdiff(goodpractice::all_checks(), "covr")
+  checks = setdiff(
+    goodpractice::all_checks(),
+    "covr")
 )
 
 devtools::check()
-# devtools::check() # from RStudio button too: non interactive session!
+# devtools::check() executed in non interactive session by gp().
 
 
 ## Before CRAN